False positives (#58)
* bug fix - macro/copy did not correctly parse lines skipped by lookahead
* Merge branch 'master' of https://github.gwd.broadcom.net/MFD/HlasmPlugin
* caching and continuing of attribute lookahead
* performance patch
* bug fix -bad macro operand parsing
* bug fix - bad text retrieval of SPACE
* bug fix - not properly evaluating substitution of value with spaces in instruction field
* some mistakes
* bug fix - bad substitution of macro param
* added sysmac sysnest
* bug fix - bad comparing in character expressions
* bug fix - X and B constants can have spaces in their definition
* macro grammar bug
* removed dangling string highlighting in lookaheaded statements
* fixed bad range and highlighting in macro reparsed statement
* fixed bad range of syntax errors in macro reparsed statements
* warnings
* missing include
* small grammar fixes, removed redundant std::move
* reworked convert switch in context_manager

+++ parser_library/src/debugging/macro_param_variable.cpp
@@ -50,7 +50,7 @@ std::vector<variable_ptr> macro_param_variable::values() const
 	}
 	else
 	{
-		for (int i = 1; i <= size(); ++i)
+		for (size_t i = 1; i <= size(); ++i)
 		{
 			child_index.back() = i;
 			vals.push_back(std::make_unique<macro_param_variable>(macro_param_, child_index));
